Fix(Oxygen) : - Ajout gestion oxygen quand le joueur meurt + vérification de l'état de la couleur du brouillard avant de prendre des dégats

+++ src/main/java/net/zed964/obscure_stars/constants/EffectsConstants.java
@@ -9,11 +9,11 @@ public class EffectsConstants {
 
     public static final float SUFFOCATION_FOG_FAR_END = 2.5F;
 
-    public static final int SUFFOCATION_FOG_COLOR_RED_END = 0;
+    public static final float SUFFOCATION_FOG_COLOR_RED_END = 0F;
 
-    public static final int SUFFOCATION_FOG_COLOR_GREEN_END = 0;
+    public static final float SUFFOCATION_FOG_COLOR_GREEN_END = 0F;
 
-    public static final int SUFFOCATION_FOG_COLOR_BLUE_END = 0;
+    public static final float SUFFOCATION_FOG_COLOR_BLUE_END = 0F;
 
     /**
      * Constructeur privé par défault
